airflow_retraining_dag.py
"""
Airflow DAG for Automated Model Retraining Pipeline
This DAG runs daily to:
1. Check for data drift
2. Retrain models if drift detected
3. Evaluate new model performance
4. Deploy if performance is acceptable
"""

from airflow import DAG
from airflow.operators.python import PythonOperator
from datetime import datetime, timedelta
import pandas as pd
import pickle
import logging

# Default arguments
default_args = {
    'owner': 'retail-pulse',
    'retries': 2,
    'retry_delay': timedelta(minutes=5),
    'start_date': datetime(2024, 1, 1),
}

# DAG definition
dag = DAG(
    'churn_model_retraining_pipeline',
    default_args=default_args,
    description='Daily churn model retraining and validation',
    schedule_interval='@daily',
    catchup=False,
)

def load_production_data():
    """Load latest production data for retraining"""
    try:
        # In production, load from database/data warehouse
        # df = pd.read_sql("SELECT * FROM retail_transactions WHERE date >= CURRENT_DATE - INTERVAL 30 DAY", conn)
        
        logging.info("✅ Production data loaded")
        return {'status': 'success', 'rows': 0}  # Placeholder
    except Exception as e:
        logging.error(f"❌ Failed to load production data: {str(e)}")
        return {'status': 'failed', 'error': str(e)}


def check_data_drift():
    """Check for statistical drift in features"""
    try:
        # Import drift detection module
        # from drift_detection import DriftDetector
        # detector = DriftDetector(reference_data, current_data)
        # drift_results = detector.detect_statistical_drift()
        
        # For now, simulated check
        drift_detected = False
        
        if drift_detected:
            logging.warning("⚠️ Data drift detected - proceeding with retraining")
        else:
            logging.info("✅ No significant drift detected")
        
        return {'drift_detected': drift_detected}
    except Exception as e:
        logging.error(f"❌ Drift detection failed: {str(e)}")
        return {'drift_detected': False, 'error': str(e)}


def retrain_churn_model(ti):
    """Retrain XGBoost churn model"""
    try:
        drift_info = ti.xcom_pull(task_ids='check_drift')
        
        if not drift_info.get('drift_detected', False):
            logging.info("⏭️ Skipping retraining (no drift)")
            return {'retrained': False, 'reason': 'no_drift'}
        
        logging.info("🔄 Starting model retraining...")
        
        # Simulate retraining
        # X_train, y_train = prepare_training_data()
        # model = xgb.XGBClassifier(**best_params)
        # model.fit(X_train, y_train)
        
        logging.info("✅ Model retrained successfully")
        return {'retrained': True, 'model_version': 'v2.0'}
    except Exception as e:
        logging.error(f"❌ Model retraining failed: {str(e)}")
        return {'retrained': False, 'error': str(e)}


def evaluate_model(ti):
    """Evaluate new model against test set"""
    try:
        retrain_info = ti.xcom_pull(task_ids='retrain_model')
        
        if not retrain_info.get('retrained', False):
            logging.info("⏭️ Skipping evaluation (no new model)")
            return {'evaluation_passed': False}
        
        logging.info("📊 Evaluating model performance...")
        
        # Simulate evaluation
        # y_test_pred = new_model.predict_proba(X_test)
        # auc = roc_auc_score(y_test, y_test_pred)
        # accuracy = accuracy_score(y_test, y_test_pred > 0.5)
        
        auc_score = 0.82  # Simulated
        min_auc = 0.78    # Minimum acceptable AUC
        
        evaluation_passed = auc_score >= min_auc
        
        if evaluation_passed:
            logging.info(f"✅ Model evaluation passed (AUC: {auc_score})")
        else:
            logging.warning(f"❌ Model evaluation failed (AUC: {auc_score} < {min_auc})")
        
        return {
            'evaluation_passed': evaluation_passed,
            'auc_score': auc_score,
            'status': 'pass' if evaluation_passed else 'fail'
        }
    except Exception as e:
        logging.error(f"❌ Model evaluation failed: {str(e)}")
        return {'evaluation_passed': False, 'error': str(e)}


def deploy_model(ti):
    """Deploy model to production if validation passed"""
    try:
        eval_info = ti.xcom_pull(task_ids='evaluate_model')
        
        if not eval_info.get('evaluation_passed', False):
            logging.info("⏭️ Skipping deployment (evaluation failed)")
            return {'deployed': False, 'reason': 'evaluation_failed'}
        
        logging.info("🚀 Deploying model to production...")
        
        # Simulate deployment
        # 1. Save model artifacts
        # 2. Update model registry
        # 3. Switch serving endpoint
        # 4. Log deployment event
        
        logging.info("✅ Model deployed successfully")
        return {'deployed': True, 'deployment_time': datetime.now().isoformat()}
    except Exception as e:
        logging.error(f"❌ Model deployment failed: {str(e)}")
        return {'deployed': False, 'error': str(e)}


def log_pipeline_results(ti):
    """Log final pipeline results"""
    try:
        load_info = ti.xcom_pull(task_ids='load_data')
        drift_info = ti.xcom_pull(task_ids='check_drift')
        retrain_info = ti.xcom_pull(task_ids='retrain_model')
        eval_info = ti.xcom_pull(task_ids='evaluate_model')
        deploy_info = ti.xcom_pull(task_ids='deploy_model')
        
        pipeline_result = {
            'timestamp': datetime.now().isoformat(),
            'data_loaded': load_info.get('status') == 'success',
            'drift_detected': drift_info.get('drift_detected', False),
            'model_retrained': retrain_info.get('retrained', False),
            'evaluation_passed': eval_info.get('evaluation_passed', False),
            'model_deployed': deploy_info.get('deployed', False),
        }
        
        logging.info(f"📋 Pipeline Summary: {pipeline_result}")
        
        # In production: Save to logging system, dashboards, alerts, etc.
        # send_slack_notification(pipeline_result)
        # save_to_database(pipeline_result)
        
        return pipeline_result
    except Exception as e:
        logging.error(f"❌ Failed to log results: {str(e)}")
        return {'error': str(e)}


# Define DAG tasks
load_data_task = PythonOperator(
    task_id='load_data',
    python_callable=load_production_data,
    dag=dag,
)

check_drift_task = PythonOperator(
    task_id='check_drift',
    python_callable=check_data_drift,
    dag=dag,
)

retrain_task = PythonOperator(
    task_id='retrain_model',
    python_callable=retrain_churn_model,
    dag=dag,
    provide_context=True,
)

evaluate_task = PythonOperator(
    task_id='evaluate_model',
    python_callable=evaluate_model,
    dag=dag,
    provide_context=True,
)

deploy_task = PythonOperator(
    task_id='deploy_model',
    python_callable=deploy_model,
    dag=dag,
    provide_context=True,
)

log_task = PythonOperator(
    task_id='log_results',
    python_callable=log_pipeline_results,
    dag=dag,
    provide_context=True,
)

# Set task dependencies
load_data_task >> check_drift_task >> retrain_task >> evaluate_task >> deploy_task >> log_task

# Task documentation
load_data_task.doc_md = """
**Load Production Data**

Load latest retail transaction data from data warehouse.
- Filters data from last 30 days
- Performs basic validation
- Logs metadata
"""

check_drift_task.doc_md = """
**Detect Data Drift**

Uses statistical tests (KS test) and Evidently AI to detect distribution shifts.
- Compares current vs baseline features
- Flags significant changes
- Logs drift metrics
"""

retrain_task.doc_md = """
**Retrain Model**

Retrains XGBoost churn model if drift detected.
- Uses Optuna for hyperparameter tuning
- Trains on latest data
- Saves model artifacts
"""

evaluate_task.doc_md = """
**Evaluate Model**

Validates new model against test set.
- Computes ROC-AUC, Accuracy, F1
- Compares against baseline model
- Decides if model is production-ready
"""

deploy_task.doc_md = """
**Deploy Model**

Deploys validated model to production.
- Updates model registry
- Switches serving endpoint
- Logs deployment event
"""

if __name__ == "__main__":
    logging.info("✅ Airflow DAG loaded successfully")
