requirements.txt
streamlit==1.37.0
pandas==2.2.0
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.4.2
xgboost==2.0.3
prophet==1.1.5
shap==0.43.8
optuna==3.4.0
statsmodels==0.14.0
scipy==1.13.0
apache-airflow==2.8.1
evidently==0.4.22
scikit-optimize==0.9.0
openpyxl==3.1.2
pyarrow==15.0.0
plotly==5.24.1
